+++ src/main/java/mei/response/DeleteTaskResponse.java
@@ -1,4 +1,4 @@
-package response;
+package mei.response;
 
 import mei.tasks.Task;
