+++ src/main/java/mei/response/MarkTaskResponse.java
@@ -1,19 +1,19 @@
-package response;
+package mei.response;
 
 import mei.tasks.Task;
 
 /**
- * Represents the response to unmark an existing user task.
+ * Represents the response to mark an existing user task.
  */
-public class UnmarkTaskResponse extends Response {
+public class MarkTaskResponse extends Response {
     private static final String[] RESPONSES = new String[] {
-        "It's alright to take things easy.",
-        "I've unchecked this task for you to revisit next time!"
+        "You've completed this? That's amazing!",
+        "I've noted down your achievement, congratulations!"
     };
 
     private Task taskToIncludeInResponses;
 
-    public UnmarkTaskResponse(Task taskToIncludeInResponses) {
+    public MarkTaskResponse(Task taskToIncludeInResponses) {
         this.taskToIncludeInResponses = taskToIncludeInResponses;
     }
