+++ src/main/java/mei/response/Response.java
@@ -1,4 +1,4 @@
-package response;
+package mei.response;
 
 import java.util.Arrays;
